Change example to users api
This change alters the example with a more useful demonstration of how
to use the auto-generated SDK. As part of the change a few hand edits
of the auto-generated code were required to get it
working. Unfortunately openapi-generator seems to create many types
with anonymous interface{} or excessive struct{} wrappers, so
temporarily work around that through manual edits.